+++ packages/site/src/pages/cards/SendHelloHessage.tsx
@@ -1,0 +1,63 @@
+/* eslint-disable no-alert */
+import { FC, useContext } from 'react';
+import { Card, SendHelloButton } from '../../components';
+import {
+  MetamaskActions,
+  MetaMaskContext,
+} from '../../contexts/MetamaskContext';
+import {
+  getCurrentNetwork,
+  sendHello,
+  shouldDisplayReconnectButton,
+} from '../../utils';
+import { validHederaChainID } from '../../utils/hedera';
+
+type Props = {
+  currentChainId: string;
+  setCurrentChainId: React.Dispatch<React.SetStateAction<string>>;
+  hederaAccountConnected: boolean;
+};
+
+const SendHelloHessage: FC<Props> = ({
+  currentChainId,
+  setCurrentChainId,
+  hederaAccountConnected,
+}) => {
+  const [state, dispatch] = useContext(MetaMaskContext);
+
+  const handleSendHelloClick = async () => {
+    try {
+      setCurrentChainId(await getCurrentNetwork());
+      await sendHello();
+    } catch (e) {
+      console.error(e);
+      dispatch({ type: MetamaskActions.SetError, payload: e });
+    }
+  };
+
+  return (validHederaChainID(currentChainId) && hederaAccountConnected) ||
+    (!validHederaChainID(currentChainId) && !hederaAccountConnected) ? (
+    <Card
+      content={{
+        title: 'Send Hello message',
+        description:
+          'Display a custom message within a confirmation screen in MetaMask.',
+        button: (
+          <SendHelloButton
+            buttonText="Send message"
+            onClick={handleSendHelloClick}
+            disabled={!state.installedSnap}
+          />
+        ),
+      }}
+      disabled={!state.installedSnap}
+      fullWidth={
+        state.isFlask &&
+        Boolean(state.installedSnap) &&
+        !shouldDisplayReconnectButton(state.installedSnap)
+      }
+    />
+  ) : null;
+};
+
+export default SendHelloHessage;
